Move unload and load to the table class instead (although view can still use unload)
